python-simulations/ocr-grading/training/requirements-training.txt
# Training Requirements for Custom OCR Models

# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Data Augmentation
albumentations==1.3.1
imgaug==0.4.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
pillow-heif==0.13.0

# Utilities
numpy==1.26.2
tqdm==4.66.1
matplotlib==3.8.2
tensorboard==2.15.1

# Dataset handling
scikit-learn==1.3.2
pandas==2.1.4

# Model evaluation
editdistance==0.6.2  # For character error rate (CER)
jiwer==3.0.3  # For word error rate (WER)

# Optional: Pretrained models
transformers==4.35.2  # For TrOCR
timm==0.9.12  # Vision models
